AP-71: Provide registry auth to docker-bench-bootstrap
The docker-bench-bootstrap may be the first time a given
node has run Apollo. As such, the image may not be present.
So, we need to provide registry auth to the service, so the
Swarm nodes can download the image.

This is currently done using the rox_registry_auth Docker
secret, which is to be mounted into the Swarm agent.