Swarm: Fix initialization order bug that led to benchmark relay panics
The Swarm agent was recently refactored to reduce copied code
between it and the new Kubernetes agent. However, this led to
the benchmark payload relayer being initialized with a nil
pointer for the gRPC client connection, because the pointer
was copied before the connection was established.

Service initialization is moved after the other initialization
steps are complete and started, so the service can have a valid
connection.