AP-427: Change cluster ZIP API to take an ID, not a spec
Change /api/extensions/clusters/zip to take a resource-by-ID object
instead of a cluster specification.

This enables the UI to create the cluster in one step and download
the ZIP in the next, and enables coming back later to download a
new ZIP if the previous one is lost.

Pull request #303
Approved-by: Connor Gorman <cgorman@stackrox.com>