Stop depending on WKT empty.Empty (stackrox/rox#44)
The explanation of why this sucks is kind of long and complicated,
but essentially this:
* gogo-proto adds new methods to the normal proto.Message interface
so that objects can marshal and unmarshal themselves using generated
code.
(see github.com/gogo/protobuf/proto/custom_gogo.go)
* We compile our protos with gogo to include those methods.
* The normal compiled versions of well known types like Timestamp
don't include those methods, so gogo ships it's own package of them
and uses aliases to map them.
* protoc-gen-grpc-gateway doesn't honor those mappings, but it doesn't
matter for anything except input types and output types, which are all
in v1 except for empty.Empty

(This works now because we simply didn't map Empty, but that causes other
problems with bazel build options we aren't using yet.)