ROX-225: Restart Sensor stream if Central disconnects (stackrox/rox#63)
If Sensor ever encountered a problem with its gRPC stream to Central,
it would never attempt to reconnect.
Change Sensor main() to reinitialize the stream if it exits.

Change pkg/sensor.Sensor to use the new concurrency library and
make it restartable.

Change the image enricher's metric registration process to allow it to
be safely created more than once.

Change Enforcer to exit when it is told to stop.