MTLS: Add SANs for [service-name].stackrox (stackrox/rox#198)
Add Subject Alternative Names with DNS names of the form
[service-name].stackrox to all issued certificates.
Change CommonName struct to represent a Subject.

This specifically allows the client to connect through an
OpenShift route that uses a Server Name Indication like
"central.stackrox", even without making central.stackrox
resolve in the remote cluster, because the gRPC endpoint
and the ServerName are configured separately.

ROX-330