Email: Don't hang if talking plaintext to a TLS server (stackrox/rox#232)
ROX-366: Fix the callsite of smtp.NewClient so that it times out in seconds
instead of minutes in certain failure cases, including when expecting
a TLS server to speak in plaintext. (This can occur if TLS is
incorrectly disabled by the user but the server is running TLS.)
Test the full truth-table of email TLS options in the QA BAT suite
to prevent regression.