Enable ReadOnlyRootFS on Central and Sensor (stackrox/rox#174)
Specify a VOLUME instruction in the Dockerfile for paths that need
to be writable, and otherwise give Sensor and Central a read-only
root file system.

Clear the search index path before trying to use it again, so that
restarts with a VOLUME work properly.

Also fix some 3-space indentation in Swarm.