CI: Fix GKE subnet size to avoid claiming /14s (stackrox/rox#340)
GKE clusters can sometimes fail to provision because of contention
for 10.0.0.0/8 in the stackrox-ci project. Reduce the size of the
subnet requests and explicitly specify subnet sizes for nodes,
pods, and services to stave off such issues.

Also, delete clusters before retrying their creation, since
the name will be the same as the previous attempt.