Try to track down grpc weirdness (stackrox/rox#596)
Sometimes, we see Central completely deadlocked. The only symptom is a log line where the forwarded request from the gateway has a delay of about 1 minute. Send SIGABRT to self at this point so we can get a stack trace dump the next time we see this. Also, since Central deadlocks whenever this happens, killing it is probably the right thing to do anyway.