Setup migrator binary (stackrox/rox#1067)
This is the first step towards running migrations before Central starts. This PR does the following:

Add a migrator binary.
Run it in the entrypoint, just before Central. If it fails, the central binary never starts.
Forbid the migrator from importing any code from the rest of the project, to ensure backwards-compatibility. This required duplicating some Bolt-related code that was in pkg.