ROX-1348 Remove polling from compliance statistics (stackrox/rox#1370)
Use apollo callbacks and component state to decide when to flush the store and re-poll for queries. Use some explicit react state to know when we should or should not be polling for job statuses.

When all jobs complete, nuke the store from orbit. This is a crappy practice, but since the apollo cache is only compliance stuff right now, the innocent bystanders are pretty few. In the near future we should switch to more selective cache invalidation or requerying selection.