ROX-1492 Automatically register scanner integration. (stackrox/rox#1493)
If scanner ping succeeds, automatically register an integration
for it, then save it to the db.

Failure to add won't get saved, so you can go add it yourself and then
see the error message. If it doesn't get saved, it'll still work which is a little
weird, but we'll still try to save it to the db after a restart.