ROX-2186 user certs take effect immediately (stackrox/rox#2584)
tls.Config isn't fully mutable (CertPool doesn't have remove, for example),
so instead we create a new one whenever things substantively change, and
we return it as the client-specific config during handshake.

Also ClientCAManager no longer really describes what this does, so move
it around.

(Also this adds -v to go generate because it was timing out in the build.)