Provision custom GCP Marketplace service account for installing into the stackrox namespace (stackrox/rox#2690)
When the Deployer image is run as part of a GCP Marketplace installation, the service account used isn't allowed to create resources in different namespaces. In the likely event that a user selects a non-stackrox namespace for installation, the product installation will fail. This PR provisions a clusterrole service account, and acts on its behalf during Deployer installation.