ROX-2735 oneof resolvers should return nil for no value (stackrox/rox#2820)
A proto oneof field can be set to any of the possible fields, or nil if unset. Because we would return an object that only checked which possible oneof field we were during the type assertion, we would end up returning an empty object to the graphql client. apollo-graphql is using type introspection for caching sanity, and would crash on those empty types.