ROX-2163 Fix long-held lock in policy set (stackrox/rox#2892)
It looks like we hold this lock for too long. There's also a lot of contention on it (many goroutines in the stacktrace are waiting on it). Instead of holding it for the entire policy evaluation period, just hold the lock long enough to clone the map. (While at it, also add some generics for map cloning.)