Fix race condition affecting persisting scans from Scanner V2 (stackrox/rox#3023)
Sometimes, after getting the scan, we wouldn't be able to save the images to the DB because we didn't have the ID. This is because we don't have the ID on image creates. This explains the symptom where we were seeing a violation that required a scan, but not seeing the corresponding scans in the Images tab.

Fix this by replacing our copy in the pendingCache with the most recent deployment object. This also ensures that we always perform our policy evaluation on the most up-to-date, and hence "correct", copy.