Define and implement Central remote control API for upgrader (stackrox/rox#3331)
Implement a remote control API in Central which the upgrader must poll after the execution of every stage. This API updates the status of the upgrade in the DB, while telling the upgrader what it should do next.

Some details:

There were just too many cases for simple if-else statements, so wrote a simple state machine in a separate package. Added unit tests to make sure that it is kept up to date if people add new workflows/stages.
Moved upgrader workflow definitions and stage names out of upgrader into pkg, to share the constants with Central.
Remove the RecordProgress endpoint from the previous PR. We don't want to allow clients to specify the end state that goes in the DB; that's for Central to determine.