ROX-3017-Add links to manually trigger sensor upgrade for UNSET and FAILED_ROLLBACK statuses (stackrox/rox#3376)
On Clusters table, add button in Upgrade status field for the following conditions:

For clusters for which and upgrade is available, but client has auto-upgrade setting disabled, make “Upgrade available” clickable, in order to trigger an upgrade on the sensor for that cluster.

For clusters where an upgrade, either auto-triggered, or manually triggered, failed, add a “Retry upgrade” button after the failure status message.

## Testing Performed

1. Automated testing: Changed unit tests for the status helper function for the statuses UNSET and UPGRADE_ERROR_ROLLED_BACK.
2. Manual testing: un-stashed my mock upgrade status code, and forced cluster response to each of those two statuses