ROX-185-add flag for creating upgrade service account (stackrox/rox#3400)
## Description

Add a checkbox to the UI on the Deployment panel of the Clusters page sidebar, before the Download button, to allow the user to choose whether or not to create an admin “upgrader” account, as part of the download package.

This setting choice is for each time the package is downloaded. It is not persisted in the UI, nor in the database, at this time.

A createUpgraderSA boolean

in the payload to POST /api/extensions/clusters/zip

along with the clusterId{ id: string, createUpgraderSA: boolean }

## Testing Performed

Manual testing

<img width="1472" alt="Screen Shot 2019-09-07 at 9 34 24 PM" src="https://user-images.githubusercontent.com/715729/64482032-945f3a80-d1b7-11e9-875f-dc5393dc0fe1.png">

Commits:
* Add toggle for creating service account for cluster

* Correct name of createUpgraderSA flag

* Update POST param for createUpgraderSA flag