Hide K8s and OpenShift tiles on Integrations page when ROX_SENSOR_AUT… (stackrox/rox#3401)
## Description

I had left it in until this point as a way to double-check, but I will do a little PR to hide it with the same flag, so only one or the other shows up.

When flag is OFF, Clusters page will not show up in subnav of Configure.
When flag is ON, Kubernetes and OpenShift tiles will not show up on Integrations page.

## Testing Performed

Manual testing

<img width="1472" alt="Screen Shot 2019-09-08 at 5 31 59 PM" src="https://user-images.githubusercontent.com/715729/64494767-fc1c9080-d25e-11e9-9189-9a1431f63a7b.png">

Commits:
* Hide K8s and OpenShift tiles on Integrations page when ROX_SENSOR_AUTOUPGRADE flag is enabled

* Address PR comments