ROX-3232 Be more helpful to users when an upgrade fails (stackrox/rox#3411)
Fix determination of upgrade states so that, if the most recent upgrade failed, we show the user why it failed.
Populate the upgrade status detail on the side panel, if it exists.
Make the clusters page own the upgrade status so that the side panel state is in sync with the parent as it polls.