ROX-3248-Fix UI issues on clusters page (stackrox/rox#3431)
## Description

Inconsistent visual treatment of toggles.

Textbox for entering monitoring end point cut off the text.

The dropdown arrow is disconnected from the select box.(when you click the dropdown arrow ,no effect)

Double-checking of dark mode for side panel (both pages in the wizard) . (cannot reproduce Connor’s screenshots from the slack channel)

ErrorErrors from the API when clicking download YAML are not surfaced to the user

There is some weird bug where the cluster type is not set to Kubernetes in the API request to the backend unless the user explicitly clicks on that select

The cluster form for new clusters has no placeholders and no default values.

The cluster form uses NO_COLLECTION as the default value for the collection field, but the default should be KERNEL_MODULE_COLLECTION

-The cluster table and side panel show indeterminate upgrade status when the upgrade status is null. The table should show a - and the panel should just not show the upgrade status.- (Fixed under different ticket: ROX-3243-multi-cluster-inconsistent-messages-bug)

When you click on the checkmark next to a cluster, it throws a “ypeError: Cannot read property 'action' of null”

After creating a cluster, clicking on the Download YAML File And Keys gives an error from the backend (see screenshots)

Set UI level feature flag for auto-upgrade to true by default. The new clusters tab is not showing when user first logs in because the feature flag API call returns 401, so the UI defaults to its feature flag defaults

## Testing Performed

Manual testing

Commits:
* Default collection field for new clusters should be KERNEL_MODULE_COLLECTION

* Fix click on the checkmark next to a cluster, it throws a “ypeError: Cannot read property action of null”

* Set UI level feature flag for auto-upgrade to true by default

* Fix Clusters form style, defaults, placeholders: 2, 3, 8, 9

* Fix error handling on Cluster form and ZIP downloading

* Fix dropdown arrow disconnected from the select box

* Make Cluster table scroll horizontally at small window widths

* Address all PR comments  but dev/prod defaults

* Adjust cluster defaults based on release status of env

* Address Viswa manual testing finds