ROX-3276 namespaces list page takes too long so remove policies from query (stackrox/rox#3447)
## Description

Latest scale measurements on config mgmt preview tag build, has the namespaces list page take ~7 seconds to load.

Noticed that the graphql query is still querying for policies violated which can be removed to significantly improve performance.

## Testing Performed

Manual testing
1. Changed the table to use only the policy overall status field from the backend to display the status column, and verified rendered the same.
2. Then, duplicated the query and removed the policies sub-list from the fields list of the duplicate. Verified that the display remained the same.

Note: I did not remove the original query from the query def file, because it appears to be used in the Compliance section, and I wanted to limit the scope of this change.

Before screenshot:
![Screen Shot 2019-09-12 at 3 36 00 PM](https://user-images.githubusercontent.com/715729/64816700-31cebb80-d576-11e9-9b56-6ac12d6bd82c.png)

After screenshot:
![Screen Shot 2019-09-12 at 3 46 35 PM](https://user-images.githubusercontent.com/715729/64816713-3bf0ba00-d576-11e9-9b03-26cf90085171.png)

Commits:
* Change namespace list policy status to use the overall status from the backend

* Remove policies list from Namespaces list query in Config Mgt