Updated config entity queries to use fewer fields (stackrox/rox#3429)
Updated the entity page single queries to follow a (hopefully) simpler pattern that uses a defaultQuery for the overview tab and a getQuery() function to generate the smallest possible query for each tab. The goals were

Minimize fields being queried
Improve readability of entity page queries
Minimize repeated logic in entity page query creation.
For that last one, I moved shared functionality into the queryService, which requires the API to be as standard as possible. If API is created that isn't consistent, we'll end up pushing back sooner.

In addition, this also fixes some bugs and inconsistencies I found along the way. For example, EntityPageHeader wasn't able to find a name for a subject.