Make it possible to QA a release up to six months after it released (stackrox/rox#3434)
Currently, we cannot run a release internally more than 30 days after it was built. This means we need to ask SEs for a prod license to do that, which seems more risky, especially given that the QA licenses are only valid for at most 16 days, and can be used only in restricted dev environments. Extend it to 180 days (seems unlikely we'll want to test a release more than six months old).