ROX-3293: Tune up styling and wording in cluster create/edit view (stackrox/rox#3476)
Add an informational/"guidance" type to the Message component
(and use it).

Move "Cluster Type" to the top of the form and mark it required,
since it is critical.

Open the Dynamic Configuration card by default.

Overhaul styling on the deployment (download+deploy) page.

Adjust a number of words/phrases for understandability and simplicity.