ROX-3472 API page doesn't load (stackrox/rox#3666)
Looks like some dependency versions got out of sync and styled-components
wasn't working right. I bumped redoc and styled-components and it still
wasn't rendering, so I rewrote the page with hooks and the problem fixed
itself, so... yay?

It is, for what it's worth, way easier to read the code than it was before.