ROX-3544-Create pattern for showing CVEs in VulnMgmt findings (stackrox/rox#3722)
## Description

Spike to find a way to embed the CVEs and/or Policies lists in the Collapsible "Findings" section of each entity's Overview tab.

Started with CVEs because that list is mostly ready for "prime time".

## Checklist

If any of these don't apply, please comment below.

## Testing Performed

Manual testing, with Deployments Overview only, for now

![Screen Shot 2019-10-18 at 10 48 42 AM](https://user-images.githubusercontent.com/715729/67117421-82e65680-f197-11e9-81e4-43e279e9f0a4.png)

Commits:
* Create pattern for showing CVEs in VulnMgmt findings

* Filter fixable CVEs for findings table on front-end

* Fix regression where tables in other sections with sub-components no longer opened

* Fix lint error