Test taking only top 5 vulns of risky entities on dashboard scatterplot (stackrox/rox#3972)
## Description

This suggestion from the bug-bash session, for trying to get a better spread of dots on the scatterplot:

> The frontend should average the top 5 for the scatterplot widget and see if it's a better spread

With this change, we are no longer averages ALL the vulnerabilities in an entity (Deployment, Namespace, Image, or Cluster).

Instead, we are only averaging the TOP 5 vulnerabilities. (Or if less than 5, the average of how many there are).

My initial observation, this just serves to shift the whole spread UP almost to the top of the graph--it doesn't really improve the situation we are trying to solve.

Screenshots of before and after, both with the small seed dataset, and a larger dataset are in the Testing section below.

## Checklist
- [ ] Investigated and inspected CI test results
- [ ] Unit test and regression tests added
- [ ] Evaluated and added CHANGELOG entry if required
- [ ] Determined and documented upgrade steps

If any of these don't apply, please comment below.

## Testing Performed

Manual testing

## Small seed dataset

### All
![averging_ALL_vulns_to_calculate_risky](https://user-images.githubusercontent.com/715729/68390172-b5e68f00-0132-11ea-9f36-de7c53c6a68e.png)

### Only Top 5
![averging_TOP_5_vulns_to_calculate_risky](https://user-images.githubusercontent.com/715729/68390199-c5fe6e80-0132-11ea-8b09-194a00e7c1fa.png)

## Larger dataset

### All
<img width="1512" alt="MORE_DATA_averging_ALL_vulns_to_calculate_risky" src="https://user-images.githubusercontent.com/715729/68390224-ceef4000-0132-11ea-8dc3-841e52ea5536.png">

### Only Top 5
<img width="1512" alt="MORE_DATA_averging_TOP_5_vulns_to_calculate_risky" src="https://user-images.githubusercontent.com/715729/68390244-dadb0200-0132-11ea-820f-35f6502be961.png">

Commits:
* Test taking only top 5 vulns of risky entities on dashboard scatterplot

* Testing pre-commit hook

* Testing pre-commit hook again

* Fix lint error

* Change language of Scatterplot hint to reflect weighting CVSS