Fix-missing-guard-against-no-vulns-in-RecentDetected-widget (stackrox/rox#4043)
## Description

Due to a build snafu, there were no entities detected by sensor.

This had the side-effect that when I visited the Vulnerability Management dashboard, the RecentlyDetectedVulnerabilities widget was not handling an empty array response correctly, and was crashing the whole app.

I fixed the bug, so that the widget now correctly displays "No vulnerabilities found" in those situations.

## Checklist
- [ ] Investigated and inspected CI test results
- [x ] Unit test and regression tests added
- ~Evaluated and added CHANGELOG entry if required~
- ~Determined and documented upgrade steps~

If any of these don't apply, please comment below.

## Testing Performed

* Unit test on the RecentlyDetectedVulnerabilities component
* Manual testing

Before:
![Screen Shot 2019-11-14 at 11 23 51 AM](https://user-images.githubusercontent.com/715729/68881688-996ec780-06db-11ea-94e9-e4facdfec167.png)

After:
![Screen Shot 2019-11-14 at 11 22 27 AM](https://user-images.githubusercontent.com/715729/68881718-a7bce380-06db-11ea-8a0c-8ec7fc75d8e1.png)