ROX-3643: Change rc-tooltip override classes to use new tooltip color (stackrox/rox#4080)
## Description

To fix bug noticed in Vuln Mgmt, https://stack-rox.atlassian.net/browse/ROX-3643

Updated the override classes for the 3rd-party rc-tooltip in app.css, to mimic the colors of the newer HoverHint tooltip.

(I could not use the tailwind classes, which use CSS variables, because those do not work with the combination of app.css generated classes and Storybook.)

## Checklist
- [ x] Investigated and inspected CI test results
- ~ Unit test and regression tests added~
- ~ Evaluated and added CHANGELOG entry if required~
- ~Determined and documented upgrade steps~

If any of these don't apply, please comment below.

## Testing Performed

Manual testing

![Screen Shot 2019-11-20 at 3 02 56 PM](https://user-images.githubusercontent.com/715729/69275447-64b2b280-0baa-11ea-86a2-660f52525144.png)