"Fix" deployment datastore unit tests (stackrox/rox#5072)
All the deployment datastore unit tests except one were failing, but we weren't catching it because we never called mockCtrl.Finish().

It looks like the failing tests don't make sense in a DackBox world since it doesn't do reindexing at the datastore level, and since AckKeysIndexed is a no-op with DackBox.