Ignore invalid entries during DB backup and restore (stackrox/rox#5115)
Upstream: stackrox/badger#5

The gist is that there's a specific kind of invalid entry in Badger which causes it to panic on GC: entries where bitValuePointer is set (which tells badger that the value is a pointer) but the value itself is empty -- essentially, the badger version of a nil pointer exception. This PR ignores such values during backup and restore. For more details, see dgraph-io/badger#1139 (comment).