Add UI for cert expiry banners for Central/Scanner/Secured Cluster (stackrox/rox#6178)
For Central and Scanner, show warning banners when the certificate is close to expiry, similar to licensing -- if expiry is in under 14 days, the banner is yellow and has a close button, but if it is under 3 days, the banner is red and cannot be closed. The banner contains a clickable link that the user can use to download the bundle.

For sensors, the warning already exists. Simply update the message to have a clickable link that the user can use to download new certs.