Add UI to apply updated certs using automatic upgrade (stackrox/rox#6206)
Implement the UI for ROX-5081.

When certs are close to expiring, show a link that allows the user to trigger cert rotation via automatic upgrade.
Clicking the link leads to a confirmation prompt.
While the cert rotation is happening, the user sees status updates (just like they do for regular automatic upgrades).
When the cert rotation is done, add text to the prompt telling the user that cert rotation was applied at so-and-so time, and that their sensor will pick up the new credentials on next restart, to prevent confusion.