ROX-5385 Fix issues relating to multiple VerifiedChains (stackrox/rox#6281)
Handle client cert auth when a TLS connection has multiple VerifiedChains.

In each verified chain, the first element should be always be identical, and equal to the leaf cert.
Therefore, when validating the token against the cert, if there are multiple VerifiedChains, we can safely pick the first one arbitrarily, since we only care about the leaf cert (specifically, that the leaf cert's fingerprint matches the one in the token).

Also use the same idea in the userpki identity extractor (which is used when client certs are presented, but without a token). In this case, it's a bit more complicated because we could technically have different auth providers in different verified chains, so it will non-deterministically pick one. This is basically what we do when we generate the token anyway, so seems better to be consistent here.

Separately, we also return an error in the mTLS service identity extractor if there are multiple VerifiedChains. Since our identity extractor list returns an error if any of the identity extractors returns an error, and the mTLS extractor runs first, control never even goes to the tokenbased service extractor (which calls backend.Validate()) when there are multiple VerifiedChains. I've fixed that to return nil, nil when there are multiple VerifiedChains, since we don't expect that with mTLS.