Make the regenerate dev license job run every day, but be smarter (stackrox/rox#6518)
Make the regenerate dev license job run every day, but be idempotent:

Only create the PR if the license is expiring in 15 days or less.
Don't create a PR if one exists already.
Also add a --json flag to roxctl central license info to print license info as JSON, for easy scripting.