Hotfix upgrade tests with a prod license (stackrox/rox#6570)
Use a prod license (added to CircleCI under ROX_PROD_LICENSE) to enable upgrade tests to still work.

Also opportunistically fix the license service to return false for whether the license is accepted unless the status is VALID or NOT_YET_VALID (following the behaviour documented in the proto file).