Update builder to use Go 1.15 (stackrox/rox#6594)
Update builder (and go.mod) to use the current go version.

Also make some improvements to the workflow:

CI now builds the builder image lazily, if it does not exist in Dockerhub. The tag of the builder image is determined by the SHA of the contents of the build directory, so that updating the Dockerfiles automatically results in a new tag.
CI will now explicitly fail with a helpful error message if the Go version in the builder image does not match the contents of EXPECTED_GO_VERSION.
Also had to update the certs used by the webhook server, which were invalid and rejected by Go 1.15 code. Took the opportunity to make it so those are generated while the test runs.