Make NetworkBaselineTest more stable (stackrox/rox#7164)
It looks like the issue was as follows: we do the connection between the server and the client in a busy loop, and our baseline window is set to just 90 seconds for tests, which means that sometimes we see the connection with a last seen timestamp that falls outside the baseline window. (Note that this won't be a problem outside tests since the baseline window is set to 1 hour, which is way more than any potential delays in us seeing network flows.)

To fix, don't connect in a busy loop, but just a finite number of times. (I tried doing exactly one wget but that was flaky too.)