Make `[< required >]` function provide default for `errorMsg`
This is for convenience because providing an error message on every
invocation would result in inconvenient and verbose code.

Note that in some earlier iteration of this code, i.e. in
https://github.com/stackrox/rox/pull/9765/files#diff-029a09f598b52f2e667a70c2d65c985fc40797868fb59cd6b6a9f061d0417498
I had no message argument to `required()` function.
However, this time I discovered that Helm has `required()` function
as well (and we use it in some files) which does take the error
message argument. Therefore I retained this string argument for less
surprises for anyone who switches between Helm and .htpl contexts.

Also note that I'm not adding tests for this function because
reproducing the template context for test seems to be quite
elaborate.