Hide `meta_test_only.go` from the release build
Had to do this because it is using `GetExampleVersion` which comes
from `set_version.go`. That file has build flags combination

```
//go:build !release || test
// +build !release test
```

that makes it invisible in release and causes errors during
`make style`.
Therefore propagated the same flags to `meta_test_only.go` in order
not to have the build error.