revert /pkg code to rely on build type
This is needed because in /pkg we can't be sure that ROX_IMAGE_FLAVOR
is provided. Plus this code is currently being called directly in roxctl,
which results in inconsitencies.

In the future this method will be removed in favor of the official
RHACS build, that will be selected by the env variable in central
and a command line argument in roxctl.