Unignore `//lint:file-ignore SA4006` for `*_easyjson.go` files
During my look at https://github.com/stackrox/stackrox/pull/269
it seemed suspicious if we still need to ignore lint warnings for
easyjson files. I tried removing them and `make staticcheck` is not
complaining about these files.

Note that we could change the makefile to be more fancy in terms of
detecting when to regenerate `*_easyjson.go` things and when not to
so that we could potentially get rid of `clean-easyjson-srcs`
dependency in `go-generated-srcs` but that's a bigger change and I'm
not too passionate to make it right now. :-P