Try stop the blob party when pushing scale-monitoring images
`build-scale-monitoring-and-mock-server` step is failing in CI
with error at `docker push` command saying
```
unknown blob
```
https://app.circleci.com/pipelines/github/stackrox/stackrox/2381/workflows/853751f3-9916-46a4-a37c-0d24204820f6/jobs/104727

It appears that only `build-scale-monitoring-and-mock-server` is
using `docker push` directly. Other pushes are done with the help of
the `./scripts/ci/push-as-manifest-list.sh` script and they don't
fail.
Using it for scale-monitoring in the hope that it makes the error go
away.