make go binaries executable
Lima manages virtual machines on macOS, which can be used as a
replacement to `Docker Desktop`. Currently, Lima contains a bug that
leads to go binaries not being executable if they have been built inside
a mounted volume (https://github.com/lima-vm/lima/issues/602).

As a workaround, add execution to the file permissions explicitly. This
should be a no-op when Lima is not used, for example on Linux systems.