ROX-10212: Disable monitoring support
If `MONITORING_SUPPORT=true`, then the deployment will include an
instance of Prometheus and Grafana, which is intended be used for
development and debugging purposes. The current CI jobs do not access
Prometheus or Grafana anymore - therefore the deployment is unnecessary
overhead and can be safely disabled.

The monitoring services can also lead to failed or flaky tests
due to CVE scanning of the container images, as was observed in ROX-10212.