Revert "Remove `clean-image` from `all-builds` target"
This reverts commit 3149711ad41787b732c6afb674e5384848af6c9b.
https://github.com/stackrox/stackrox/pull/1875

The reason is that `create-bundle.sh` looks not only in
`image/rhel` directory, but also takes files from `image/bin`,
`image/static-bin` and others.

As a result of the change, Makefile did not detect that bundle
needs to be regenerated when, for example, new `roxctl` binaries
were built, and old binaries were packaged in the bundle tarball
and then in image.

See also https://srox.slack.com/archives/CELUQKESC/p1654264051403859