Try ensure logs are collected from `qa-splunk-violation` namespace
The test method formerly created and deleted the namespace.
Deletion was done in `cleanup:` section which runs even on exception.

We have an interceptor that launches `collect-qa-service-logs.sh`
scripts on exception from the test method. However, the namespace
was already deleted by `cleanup:` section and so there was nothing to
collect.

A pattern that's used in other tests is to create a namespace in
`setupSpec` and delete it in `cleanupSpec`. This way the namespace
will stay for test failure interceptor to collect logs.