Respect `node-role.kubernetes.io/control-plane` label
in addition to `node-role.kubernetes.io/master`.
As was pointed out by @vikin91 in PR comments, `*/control-plane`
should be used instead of `*/master` on newer cluster versions.
See:
* https://kubernetes.io/docs/reference/setup-tools/kubeadm/implementation-details/#mark-the-node-as-control-plane
* https://github.com/kubernetes/kubernetes/blob/master/CHANGELOG/CHANGELOG-1.20.md?rgh-link-date=2022-06-28T11%3A16%3A54Z#no-really-you-must-read-this-before-you-upgrade

Note that I HAVE NOT touched Compliance Operator. It still has some
mentions of `*/master` in code. I don't understand it well enough to
be able to fearlessly change it. Maybe that's something for another
time.