ROX-11456: Add RHACS data plane dashboard (#2225)
The dashboard includes both go, gRPC and database metrics for Central. I
kept the original dashboard in `central.json` because it also includes
some Sensor metrics, which are omitted from the new dashboard.

The dashboard will be added to
https://github.com/stackrox/rhacs-observability-resources as well,
however I found it easier to develop the dashboard with the local
monitoring stack.

I also aligned the deployed Grafana version with the version deployed by
the Grafana operator - it is the latest 7.x version.