(1) unified store interface for deployments, pods and netpols
At this point the idea is to simply have a struct that holds all the stores in
a centralized manner. This will help any component that needs to access the
in-memory data to do so through this struct. Plus it helps with mocking in
tests.

The problem here is that some stores (e.g. Namespace and RBAC) are not abstracted
in the same way (e.g. returning storage objects and providing an interface). If
we're going down this path of creating a "Store" component, that we need to make
sure that all store lifecycles are managed together.