(4) Additional test for Deployment -> Network Policy association
This tests made me realize something: there is a coupling between
adding things to the store and to the graph.

I'm assuming that adding things to the resources store is done
outside of the dependency graph package. And since we need to
use the stores `Find()` functions to find the matching resources
whenever a new resource is added, this creates an odd case:

What happens if we generate a snapshot for a new resource (d1) and
when we run `Find()` on NetworkPolicy store we find a netpol (n1)
that isn't in the graph? This could happen if the caller didn't call
`GenerateSnapshotFromUpsert` for n1. In this case, should we then
add n1 to the graph and return the snapshot of d1 *with* n1? Or
should it be ignored since the dependency graph doesn't have it?