Ensure `REACT_APP_ROX_PRODUCT_BRANDING` is set for UI build in CI
from `ROX_PRODUCT_BRANDING`.

The idea is to use the `ROX_PRODUCT_BRANDING` as the primary one and
set `REACT_APP_ROX_PRODUCT_BRANDING` to the equal value. The former
is used for the back-end while the latter for UI.

We already use `ROX_PRODUCT_BRANDING` as primary in other places like
deriving image flavor in CI.

This should address the issue reported here
https://srox.slack.com/archives/CELUQKESC/p1657585769140049
with follow-on conversation here
https://srox.slack.com/archives/CELUQKESC/p1657586625489199