(6) Implement adjencency list graph and edge finder
Ok this is big one.

This is a implementation of how the depedency graph and the snapshot
tree can be implemented in a generic way. It uses an adjecency list
data structure to index resource ids to their edges (parents and
children).

Canonical ID: My insight when trying to implement this is that we
can come up with a unique id that identifies any resource in the
dependency graph. This is the canonical id. It is constructed as
such: "<kind>#<namespace>#<id>". This way we have a single string
value that can be used as a map index for any node in the map.
The identifiers structure is just a wrapper around the desconstructed
canonical id. Its name could probably be better, but for now it
does the trick.

Finders: Another abstraction that came up is the `Finder` interface.
The idea is to use the underlying store components to find the
dependencies orignally, i.e. they can implement specific methods
to search for resources based on their properties. Like `Find` in
NetworkPolicyStore or `GetMatchingDeployments` in DeploymentStore.
A finder will simply be a wrapper for each store based on the kind
and the direction of relationship we're looking for. A finder is
able to find both dependencies and dependants of the resource
requested.

Graph Construction: as mentioned in the previous commits, the
idea for building the graph was to simply unify upsert and snapshot
generation. This raised the problem of maybe finding resource
relations that are not yet in the graph. The way this was
implemented now was to recursively try to add the node to the graph
if it's not there, and to use a `StringSet` for the edges. This
way if the node we're creating the edge to is already present,
we simply create an edge in both the new node and the existing one.
If the node is not yet created, we first create it and try to find
all edges of that node. Only then we add the edges between the
two nodes.

Next step would be to implement how to handle a change in the edges.