(7) Implement edge update if network policy selector changed
This is the logic that I wanted to test from the beginning of this
exercise.

The idea is as follows:
- Given the dependency graph already know about the relationship
between two resources (r1 and r2).
- Given that we receive an update (in r2) that removes the edge between
r1 and r2 and adds one between r2 and r3.
- Then the dependency graph identifies which edges were removed
(r1-r2) and updates the graph to the new form

OLD                          NEW
[r1] ---> [r2]               [r1] -x-> [r2]

[r3]                         [r3] ---> [r2]

- Then we generate snapshots from the union between:
	- All the parents of the updated node (r2)
	- All the parents of removed edges nodes (r1)
- The final snapshot will contain both the r1 segment and the
r3 segment, such as:

Snapshot1:
[r1]
Snapshot2:
[r3] --> [r2]

This way sensor will know that it needs to reprocess both the r1
tree and the newly formed r3 tree.