Prefactoring: work around a kuttl issue with kubeconfig when running in a pod.
Normally, when running commands and scripts as part of a step, `kuttl` writes a
`kubeconfig` file and sets $KUBECONFIG in the environment to point to it.

However it seems that when it is running in a k8s pod it gets confused and
points $KUBECONFIG at an inexistent file.

So we smuggle the actual $KUBECONFIG behind kuttl's back in a separate
environment variable to any command that needs it.